- created calendar layout for CalendarWithInfoScreen - extracted common calendar components into UI elements and utilities
