+++ app/src/main/java/com/example/siyai_front_android/presentation/main/bottom_nav_container/MainNavHost.kt
@@ -58,7 +58,9 @@ fun MainNavHost(
             )
         }
         composable<MainRoute.CalendarWithInfo> {
-            CalendarWithInfoScreen()
+            CalendarWithInfoScreen(
+                viewmodelFactory = viewModelFactory
+            )
         }
         composable<MainRoute.Profile> {
             ProfileScreen(
