+++ app/src/main/java/com/example/siyai_front_android/ui/components/calendar/MultiRangeDatePicker.kt
@@ -1,24 +1,12 @@
 package com.example.siyai_front_android.ui.components.calendar
 
 import android.content.Context
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -26,22 +14,18 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.RectangleShape
-import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.siyai_front_android.R
 import com.example.siyai_front_android.domain.dto.Cycle
-import java.text.SimpleDateFormat
-import java.util.Calendar
-import java.util.Locale
+import com.example.siyai_front_android.utils.MonthData
+import com.example.siyai_front_android.utils.createMonthData
+import com.example.siyai_front_android.utils.getDateToRangeMap
+import com.example.siyai_front_android.utils.getDaysStates
 
 @Composable
 fun MultiRangeDatePicker(
@@ -121,143 +105,6 @@ private fun MonthCalendar(
     }
 }
 
-@Composable
-private fun CalendarGrid(
-    monthData: MonthData,
-    daysStates: List<DayState>,
-    onDateClick: (Long) -> Unit
-) {
-    val totalCells = monthData.startDayOfWeek + monthData.daysInMonth
-    val weeks = (totalCells + 6) / 7
-
-    BoxWithConstraints(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp)
-    ) {
-        val cellSize = maxWidth / 7
-
-        Column {
-            repeat(weeks) { weekIndex ->
-                Row(
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    repeat(7) { dayOfWeek ->
-                        val cellIndex = weekIndex * 7 + dayOfWeek
-                        val cellModifier = Modifier.size(cellSize)
-
-                        when {
-                            cellIndex < monthData.startDayOfWeek -> {
-                                Spacer(modifier = cellModifier)
-                            }
-                            else -> {
-                                val dayIndex = cellIndex - monthData.startDayOfWeek
-                                if (dayIndex < monthData.daysInMonth) {
-                                    val dayState = daysStates[dayIndex]
-                                    CalendarDay(
-                                        modifier = cellModifier,
-                                        dayState = dayState,
-                                        dayOfWeek = dayOfWeek,
-                                        daysInMonth = monthData.daysInMonth,
-                                        onDateClick = onDateClick
-                                    )
-                                } else {
-                                    Spacer(modifier = cellModifier)
-                                }
-                            }
-                        }
-                    }
-                }
-                if (weekIndex < weeks - 1) {
-                    Spacer(modifier = Modifier.height(8.dp))
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun CalendarDay(
-    modifier: Modifier,
-    dayState: DayState,
-    dayOfWeek: Int,
-    daysInMonth: Int,
-    onDateClick: (Long) -> Unit
-) {
-    val (backgroundColor, shape, textColor) = remember(
-        dayState, dayOfWeek, daysInMonth
-    ) {
-        getCellStyles(dayState, dayOfWeek, daysInMonth)
-    }
-
-    val finalModifier = remember(modifier, dayState.isToday) {
-        if (dayState.isToday)
-            modifier.border(width = 1.dp, color = Color.White, shape = CircleShape)
-        else
-            modifier
-    }
-
-    Box(
-        modifier = finalModifier
-            .background(backgroundColor, shape)
-            .clip(CircleShape)
-            .clickable(enabled = dayState.isAvailable) {
-                onDateClick(dayState.dateMillis)
-            },
-        contentAlignment = Alignment.TopCenter
-    ) {
-        Text(
-            modifier = Modifier.padding(top = 8.dp),
-            text = dayState.day.toString(),
-            color = textColor,
-            style = MaterialTheme.typography.bodyLarge
-        )
-
-        if ((dayState.rangePosition != RangePosition.None || dayState.isTempStart) && dayState.isAvailable) {
-            Box(
-                modifier = Modifier
-                    .padding(bottom = 8.dp)
-                    .size(4.dp)
-                    .background(Color.White, CircleShape)
-                    .align(Alignment.BottomCenter)
-            )
-        }
-    }
-}
-
-@Composable
-private fun RowDays(
-    modifier: Modifier = Modifier
-) {
-    val context = LocalContext.current
-    val dayHeaders = remember {
-        listOf(
-            context.getString(R.string.day_mon),
-            context.getString(R.string.day_tue),
-            context.getString(R.string.day_wed),
-            context.getString(R.string.day_thu),
-            context.getString(R.string.day_fri),
-            context.getString(R.string.day_sat),
-            context.getString(R.string.day_sun),
-        )
-    }
-
-    Row(
-        modifier = modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceEvenly
-    ) {
-        dayHeaders.forEach { day ->
-            Text(
-                text = day,
-                color = Color.White.copy(alpha = 0.6f),
-                style = MaterialTheme.typography.bodySmall,
-                modifier = Modifier.weight(1f),
-                textAlign = TextAlign.Center
-            )
-        }
-    }
-}
-
 private fun handleDateClick(
     date: Long,
     selectedRanges: List<Cycle>,
@@ -332,151 +179,9 @@ private fun processHandleClick(
     }
 }
 
-private fun getCellStyles(
-    dayState: DayState,
-    dayOfWeek: Int,
-    daysInMonth: Int
-): Triple<Color, Shape, Color> {
-    val backgroundColor = when {
-        !dayState.isAvailable -> Color.Transparent
-        dayState.isTempStart || dayState.rangePosition != RangePosition.None -> Color(0xFFE54E7C)
-        else -> Color.Transparent
-    }
-
-    val shape = when (dayState.rangePosition) {
-        RangePosition.Start -> {
-            if (dayOfWeek == 6 || dayState.day == daysInMonth) CircleShape
-            else RoundedCornerShape(topStart = 50.dp, bottomStart = 50.dp)
-        }
-        RangePosition.End -> {
-            if (dayOfWeek == 0 || dayState.day == 1) CircleShape
-            else RoundedCornerShape(topEnd = 50.dp, bottomEnd = 50.dp)
-        }
-        RangePosition.Middle -> {
-            when (dayOfWeek) {
-                0 -> {
-                    if (dayState.day == daysInMonth) CircleShape else RoundedCornerShape(
-                        topStart = 50.dp,
-                        bottomStart = 50.dp
-                    )
-                }
-                6 -> {
-                    if (dayState.day == 1) CircleShape
-                    else RoundedCornerShape(topEnd = 50.dp, bottomEnd = 50.dp)
-                }
-                else -> {
-                    when (dayState.day) {
-                        1 -> RoundedCornerShape(topStart = 50.dp, bottomStart = 50.dp)
-                        daysInMonth -> RoundedCornerShape(topEnd = 50.dp, bottomEnd = 50.dp)
-                        else -> RectangleShape
-                    }
-                }
-            }
-        }
-        else -> {
-            RoundedCornerShape(topStart = 50.dp, bottomStart = 50.dp)
-        }
-    }
-
-    val textColor = when {
-        !dayState.isAvailable -> Color.White.copy(alpha = 0.3f)
-        dayState.rangePosition != RangePosition.None || dayState.isTempStart -> Color.White
-        else -> Color.White.copy(alpha = 0.8f)
-    }
-
-    return Triple(backgroundColor, shape, textColor)
-}
-
-private fun getDaysStates(
-    monthData: MonthData,
-    dateToRangeMap: Map<Long, Pair<Int, Cycle>>,
-    tempStartDate: Long?
-): List<DayState> {
-    val today = getTodayTimeInMillis()
-
-    return (1..monthData.daysInMonth).map { day ->
-        val dateMillis = monthData.firstDayMillis + (day - 1) * 24 * 60 * 60 * 1000L
-        val rangeInfo = dateToRangeMap[dateMillis]
-        val isInRange = rangeInfo != null
-        val isRangeStart = rangeInfo?.second?.start == dateMillis
-        val isRangeEnd = rangeInfo?.second?.end == dateMillis
-        val isTempStart = tempStartDate == dateMillis
-        val isToday = dateMillis == today
-
-        val rangePosition = when {
-            isRangeStart -> RangePosition.Start
-            isRangeEnd -> RangePosition.End
-            isInRange -> RangePosition.Middle
-            else -> RangePosition.None
-        }
-
-        DayState(
-            day = day,
-            dateMillis = dateMillis,
-            isAvailable = dateMillis <= today,
-            isToday = isToday,
-            rangePosition = rangePosition,
-            isTempStart = isTempStart,
-            isInSelectedRange = isInRange,
-            rangeIndex = rangeInfo?.first ?: -1
-        )
-    }
-}
-
-private fun getDateToRangeMap(selectedRanges: List<Cycle>): Map<Long, Pair<Int, Cycle>> =
-    buildMap {
-        selectedRanges.forEachIndexed { index, range ->
-            var current = range.start
-            while (current <= range.end) {
-                put(current, index to range)
-                current += 24 * 60 * 60 * 1000
-            }
-        }
-    }
-
 private fun getMonthsData(context: Context, monthsCount: Int): List<MonthData> {
-    val currentMonth = Calendar.getInstance().apply {
-        set(Calendar.DAY_OF_MONTH, 1)
-        set(Calendar.HOUR_OF_DAY, 0)
-        set(Calendar.MINUTE, 0)
-        set(Calendar.SECOND, 0)
-        set(Calendar.MILLISECOND, 0)
-    }
-    val dateFormatter = SimpleDateFormat(
-        context.getString(R.string.date_header_format), Locale(context.getString(R.string.locale_ru))
-    )
-    val currentCalendar = Calendar.getInstance()
-
-    return (0..monthsCount).map { index ->
-        val monthOffset = -index
-        val monthCalendar = Calendar.getInstance().apply {
-            timeInMillis = currentMonth.timeInMillis
-            add(Calendar.MONTH, monthOffset)
-        }
-
-        val isCurrentMonth =
-            monthCalendar.get(Calendar.YEAR) == currentCalendar.get(Calendar.YEAR) &&
-                    monthCalendar.get(Calendar.MONTH) == currentCalendar.get(Calendar.MONTH)
-
-        MonthData(
-            year = monthCalendar.get(Calendar.YEAR),
-            month = monthCalendar.get(Calendar.MONTH),
-            monthName = dateFormatter.format(monthCalendar.time),
-            daysInMonth = monthCalendar.getActualMaximum(Calendar.DAY_OF_MONTH),
-            startDayOfWeek = (monthCalendar.get(Calendar.DAY_OF_WEEK) - 2 + 7) % 7,
-            isCurrentMonth = isCurrentMonth,
-            firstDayMillis = monthCalendar.timeInMillis
-        )
-    }
+    return (0..monthsCount).map { index -> createMonthData(context, -index) }
 }
-
-private fun getTodayTimeInMillis(): Long =
-    Calendar.getInstance().apply {
-        set(Calendar.HOUR_OF_DAY, 0)
-        set(Calendar.MINUTE, 0)
-        set(Calendar.SECOND, 0)
-        set(Calendar.MILLISECOND, 0)
-    }.timeInMillis
 
 private fun validateDateSelection(
     start: Long,
@@ -511,30 +216,7 @@ private fun ValidationError.getErrorMessage(context: Context): String = when (th
     ValidationError.MAX_RANGES_REACHED -> context.getString(R.string.error_max_reached)
 }
 
-data class MonthData(
-    val year: Int,
-    val month: Int,
-    val monthName: String,
-    val daysInMonth: Int,
-    val startDayOfWeek: Int,
-    val isCurrentMonth: Boolean,
-    val firstDayMillis: Long
-)
-
-data class DayState(
-    val day: Int,
-    val dateMillis: Long,
-    val isAvailable: Boolean,
-    val isToday: Boolean,
-    val rangePosition: RangePosition,
-    val isTempStart: Boolean,
-    val isInSelectedRange: Boolean = false,
-    val rangeIndex: Int = -1
-)
-
-enum class RangePosition { None, Start, Middle, End }
-
-enum class ValidationError {
+private enum class ValidationError {
     SINGLE_DAY_NOT_ALLOWED,
     BACKWARD_SELECTION,
     RANGES_OVERLAP,
@@ -542,7 +224,7 @@ enum class ValidationError {
     MAX_RANGES_REACHED
 }
 
-sealed class DateClickResult {
+private sealed class DateClickResult {
     data object Success : DateClickResult()
     data class Error(val validationError: ValidationError) : DateClickResult()
     data class ShowDialog(val index: Int) : DateClickResult()
